[#8] Fixed build failing if empty createContext arg provided
When running `gatsby build`, the Gatsby bundle would fail to build because all props being used by the context API's consumer were being recieved as undefined.

The specific error was "WebpackError: TypeError: Cannot read property 'darkMode' of undefined"

This is likely because the createContext() function was being exported before the globalState was actually being set. To fix this, an initial state was created to allow for a default prop value to be used by the consumers, while the remainder of the global-state-context-provider.js was being built. With this existing implementation, all state will be declared in the defaultState and globalState variables for consistency.

NOTE:
- This issue seems to be Gatsby specific and is not reproducible via CRA
- See gatsbyjs/gatsby#19255 for more detail