[#8] User site theme preference saved in localStorage
When the user reloads the website or visits it again, their previous theme (i.e. dark or light) will be used. This is a slight quality of life improvement as the user will not have to keep toggling the theme each time they re-visit the site.

Aside: React.createContext() no longer requires an initial state (don't know why), so defaultState was removed.