Favicon now fetched over HTTPS instead of HTTP
The favicon was being fetched over HTTP. There were multiple repeat calls in header.html, the last one taking precedence (which was HTTP, not HTTPS). The favicon is now fetched via the site.baseurl which is running over HTTPS (through GH pages). The site should now be "secure" and display the green lock in the address bar.