StyledComponents styles are compiled when site is built and served
Importing the normal StyledComponents library (like your do with CRA) does not work as I originally expected with Gatsby.

When running the site in development mode, the styles created would be shown correctly, but when the site was built via 'gatsby build', the index files would be unable to locate the appropriate styles since they were not being compiled with the rest of the assets. Adding the gatsby-plugin-styled-components plugin allows Gatsby to also compile the StyledComponent styles when building the site, ultimately packaging it with all other assets and files. Now the devlopment and the built sites should be identical.